Refactor codebase and consolidate downloads
The commit encompasses major refactorization of the code, primarily unifying all downloads related functionality into one module. It revises routing, optimizes database transaction logic and renames several modules and functions for code readability. By reorganizing the code, it simplifies management and future updates. All download entities are now consolidated under a 'downloads' module, with improvements designed to enhance overall application performance and maintainability.
